requirements.txt
black==23.3.0
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
requests==2.31.0

